Sync upstream ROSE 2023-2025 (#488)
## Summary
- Upstream ROSE sync (2023-10-30 → 2025-12-29) with REX-specific drop
list enforced.
- Aligns CFE/ROSE compile-only behavior and removes local compileOnly
workarounds.
- Removes dropped components (e.g., EDG/Java/BinaryAnalysis/Jovial/Ada,
ROSE-LLVM connection) and keeps CMake-first workflow.

## Key changes
- Frontend: propagate project -c to SgFile at construction; remove local
compileOnly OR logic; keep upstream-style conditionals.
- Unparser: fixes for class/typedef definition suppression and Clang
restrict handling.
- Fortran: OFP invocation fixes; Flang experimental support sync
additions (without altering third-party flang-parser).
- Build system + docs: CMake updates, Doxygen config additions, upstream
sync guide + commit log.

## Tests
- cmake --build build -j16
- ctest --test-dir build --output-on-failure -R "rex|astInterface" -j16

## Notes
- Flang driver -c propagation inconsistency tracked in
https://github.com/ouankou/rex/issues/487 (no behavior change in this PR
beyond ROSE-side compileOnly propagation).

+++ src/frontend/Experimental_General_Language_Support/Tokens.h
@@ -82,7 +82,7 @@ public:
 
 private:
   std::vector<Token> tokens_;
-  int next_;
+  size_t next_;
 
   int getTokenElement(std::istream &, std::string &);
   int getTokenComment(std::istream &, std::string &);
